+++ CheckIP.Mobile/MAUI/CheckIPApp/AppShell.xaml.cs
@@ -1,0 +1,9 @@
+namespace CheckIPApp;
+
+public partial class AppShell : Shell
+{
+	public AppShell()
+	{
+		InitializeComponent();
+	}
+}