+++ CheckIP.Mobile/MAUI/CheckIPApp/Platforms/MacCatalyst/AppDelegate.cs
@@ -1,0 +1,9 @@
+﻿using Foundation;
+
+namespace CheckIPApp;
+
+[Register("AppDelegate")]
+public class AppDelegate : MauiUIApplicationDelegate
+{
+	protected override MauiApp CreateMauiApp() => MauiProgram.CreateMauiApp();
+}