+++ CheckIP.Mobile/Xamarin/CheckIPApp/CheckIPApp/AppShell.xaml.cs
@@ -1,0 +1,17 @@
+﻿using CheckIPApp.ViewModels;
+using CheckIPApp.Views;
+using System;
+using System.Collections.Generic;
+using Xamarin.Forms;
+
+namespace CheckIPApp
+{
+    public partial class AppShell : Xamarin.Forms.Shell
+    {
+        public AppShell()
+        {
+            InitializeComponent();
+        }
+
+    }
+}