+++ CheckIP.Mobile/Xamarin/CheckIPApp/CheckIPApp/ViewModels/AboutViewModel.cs
@@ -1,0 +1,20 @@
+﻿using CheckIPApp.Services;
+using Sharpnado.MaterialFrame;
+using System;
+using System.Windows.Input;
+using Xamarin.Essentials;
+using Xamarin.Forms;
+
+namespace CheckIPApp.ViewModels
+{
+    public class AboutViewModel : BaseViewModel
+    {
+        public AboutViewModel()
+        {
+            Title = "About";
+            OpenWebCommand = new Command(async () => await Browser.OpenAsync("https://github.com/valnoxy/checkip"));
+        }
+
+        public ICommand OpenWebCommand { get; }
+    }
+}